Enable the caches.
To do this, we also need to enable the MMU.  This change attempts to add
the absolute bare minimum functionality required to bring up the ARMv7
MMU, configure an identity-mapped page-table with all physical memory
marked as cacheable, and enable the caches.

It also adds some cache management primitives in anticipation of their
necessity in making DMA drivers work.